Move tx-status listener to block-sync-renderer folder
